refactor

+++ src/App.js
@@ -3,14 +3,17 @@ import { ThemeProvider } from 'styled-components';
 import { GlobalStyle } from './globalStyles';
 import theme from './theme/theme';
 
-import Header from './components/molecules/Header';
-import CustomMeal from './components/molecules/CustomMeal';
-import Sum from './components/molecules/Sum';
-import Meal from './components/organisms/Meal';
+import { Header } from './components/molecules/Header/Header';
+import { CustomMeal } from './components/organisms/CustomMeal/CustomMeal';
+import { Sum } from './components/organisms/Sum/Sum';
+import { Meal } from './components/organisms/Meal/Meal';
+
+import { DEFAULT_MEALS } from './constants';
+import { getLocalStorage } from './helpers';
 
 const App = () => {
   const [mealsList, setMealsList] = useState(
-    JSON.parse(localStorage.getItem('mealsList')) || ['breakfast', 'lunch', 'dinner']
+    JSON.parse(localStorage.getItem('mealsList')) || DEFAULT_MEALS
   );
 
   useEffect(() => {
@@ -20,17 +23,16 @@ const App = () => {
   const [total, setTotal] = useState([]);
 
   useEffect(() => {
-    const date = JSON.parse(localStorage.getItem('date'));
+    const date = getLocalStorage('date');
     const newDate = new Date();
 
     const [day, month, year] = [newDate.getDate(), newDate.getMonth() + 1, newDate.getFullYear()];
 
-    if (date) {
-      if (day > date.day || month > date.month || year > date.year) {
-        setMealsList(['breakfast', 'lunch', 'dinner']);
-        localStorage.clear();
-      }
+    if (date && (day > date.day || month > date.month || year > date.year)) {
+      setMealsList(DEFAULT_MEALS);
+      localStorage.clear();
     }
+
     localStorage.setItem('date', JSON.stringify({ day, month, year }));
   }, []);
